# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/platform/tenants/{tenantId}/outreach-config": {
      "get": {
        "description": "Devolve ids de catálogo, slotBindings, knobs de envio e relações mínimas dos templates (id, name, language, status).",
        "operationId": "OutreachConfigController_get",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter outreach config do tenant (404 se não houver)",
        "tags": [
          "Platform — Outreach Config"
        ]
      },
      "put": {
        "description": "Cria a config se não houver row. Se já existir, 403 — use PATCH para preço. Body completo permitido no pontapé (inclui campos tenant-owned). Campos legado (outreachTemplateName, notifyTenantTemplateName, outreachContactText, headerImageUrl) retornam 400. enabled=true exige tenant ativo com phone, templates APPROVED e cobertura de slots.",
        "operationId": "OutreachConfigController_upsert",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertOutreachConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bootstrap de outreach config (somente se ainda não existir)",
        "tags": [
          "Platform — Outreach Config"
        ]
      },
      "patch": {
        "description": "Somente campos de plataforma. Campos tenant-owned (enabled, schedule, categories, leadsPerRun, etc.) no body → 403. Campos legado retornam 400.",
        "operationId": "OutreachConfigController_patch",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchPlatformOutreachConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Patch de preço (costPerLead / cashbackOnReply)",
        "tags": [
          "Platform — Outreach Config"
        ]
      }
    },
    "/tenant/{tenantId}/outreach-config": {
      "get": {
        "description": "Devolve a row completa, inclusive costPerLead/cashbackOnReply. Admin não altera preço neste path.",
        "operationId": "TenantOutreachConfigController_get",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter outreach config do tenant (inclui preço read-only)",
        "tags": [
          "Tenant — Outreach Config"
        ]
      },
      "put": {
        "description": "Cria com costPerLead=0 e cashbackOnReply=0. Se já existir, 403 — use PATCH. costPerLead/cashbackOnReply no body → 403. Super Admin só no pontapé (recurso ausente).",
        "operationId": "TenantOutreachConfigController_create",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTenantOutreachConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar outreach config se ainda não existir",
        "tags": [
          "Tenant — Outreach Config"
        ]
      },
      "patch": {
        "description": "Não aceita costPerLead/cashbackOnReply (403). Super Admin só dentro da janela de 30 min. enabled=true exige phone, tenant ativo, templates APPROVED, grants e bindings.",
        "operationId": "TenantOutreachConfigController_patch",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchOutreachConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Patch operacional (knobs, schedule, templates, enabled)",
        "tags": [
          "Tenant — Outreach Config"
        ]
      }
    },
...
    "/platform/platform-job-schedules": {
      "get": {
        "operationId": "PlatformJobSchedulesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar schedules de jobs da plataforma (sync templates e scrape)",
        "tags": [
          "Platform — Job Schedules"
        ]
      }
    },
    "/platform/platform-job-schedules/{jobKey}": {
      "get": {
        "operationId": "PlatformJobSchedulesController_getByJobKey",
        "parameters": [
          {
            "name": "jobKey",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "WHATSAPP_TEMPLATE_SYNC",
                "SCRAPE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter schedule por jobKey",
        "tags": [
          "Platform — Job Schedules"
        ]
      },
      "put": {
        "description": "jobKey deve ser WHATSAPP_TEMPLATE_SYNC ou SCRAPE. Workers fazem poll; PUT não notifica processos.",
        "operationId": "PlatformJobSchedulesController_upsert",
        "parameters": [
          {
            "name": "jobKey",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "WHATSAPP_TEMPLATE_SYNC",
                "SCRAPE"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPlatformJobScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar (upsert) cron, timezone e enabled de um job da plataforma",
        "tags": [
          "Platform — Job Schedules"
        ]
      }
    },
...
    "/platform/ops/summary": {
      "get": {
        "description": "Contadores: total de tenants, ativos, com outreach enabled, e leads no pool global (não soft-deleted).",
        "operationId": "OpsController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resumo operacional da plataforma",
        "tags": [
          "Platform — Ops"
        ]
      }
    },
    "/platform/tenants/{tenantId}/leads/stats": {
      "get": {
        "description": "Counts onde cada booleano de TenantLead é true. 404 se o tenant não existir.",
        "operationId": "OpsController_leadsStats",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantLeadStatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Funil de TenantLead por tenant",
        "tags": [
          "Platform — Ops"
        ]
      }
    },
    "/platform/leads/count": {
      "get": {
        "description": "Lead com deletedAt = null.",
        "operationId": "OpsController_leadsCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadCountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Contagem do pool global de leads",
        "tags": [
          "Platform — Ops"
        ]
      }
    },
    "/tenant/{tenantId}/leads/stats": {
      "get": {
        "description": "Counts onde cada booleano de TenantLead é true. 404 se o tenant não existir.",
        "operationId": "TenantLeadsStatsController_leadsStats",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantLeadStatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Funil de TenantLead do próprio tenant",
        "tags": [
          "Tenant — Ops"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gestão de Leads",
    "description": "API para gestão de leads",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SlotBindingsDto": {
        "type": "object",
        "properties": {
          "outreach": {
            "type": "object",
            "description": "Bindings do template de outreach, keyed por slot",
            "example": {
              "body.1": {
                "type": "literal",
                "value": "Gladson Teixeira"
              }
            }
          },
          "notify": {
            "type": "object",
            "description": "Bindings do template de notify, keyed por slot",
            "example": {
              "body.customer_name": {
                "type": "lead.name"
              }
            }
          }
        },
        "required": [
          "outreach",
          "notify"
        ]
      },
      "UpsertOutreachConfigDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Habilitar outreach",
            "default": false
          },
          "costPerLead": {
            "type": "number",
            "description": "Custo em coins por lead contatado",
            "example": 0.35
          },
          "cashbackOnReply": {
            "type": "number",
            "description": "Cashback ao receber reply",
            "default": 0
          },
          "outreachTemplateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de outreach no catálogo"
          },
          "notifyTemplateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de notify no catálogo"
          },
          "slotBindings": {
            "$ref": "#/components/schemas/SlotBindingsDto"
          },
          "schedule": {
            "type": "object",
            "description": "Mapa dia-da-semana → horas UTC",
            "example": {
              "2": [
                18
              ],
              "3": [
                18
              ],
              "4": [
                13,
                18
              ]
            }
          },
          "categories": {
            "minItems": 1,
            "description": "Categorias de leads",
            "example": [
              "Construtoras",
              "Clínicas médicas"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "leadsPerRun": {
            "type": "number",
            "minimum": 1,
            "description": "Leads contactados por execução (default 5)",
            "example": 5,
            "default": 5
          },
          "sendIntervalSeconds": {
            "type": "number",
            "minimum": 0,
            "description": "Intervalo em segundos entre envios (default 5)",
            "example": 5,
            "default": 5
          }
        },
        "required": [
          "costPerLead",
          "outreachTemplateId",
          "notifyTemplateId",
          "slotBindings",
          "schedule",
          "categories"
        ]
      },
      "PatchPlatformOutreachConfigDto": {
        "type": "object",
        "properties": {
          "costPerLead": {
            "type": "number",
            "description": "Custo em coins por lead contatado"
          },
          "cashbackOnReply": {
            "type": "number",
            "description": "Cashback ao receber reply"
          }
        }
      },
      "UpsertTenantOutreachConfigDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Habilitar outreach",
            "default": false
          },
          "outreachTemplateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de outreach no catálogo"
          },
          "notifyTemplateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de notify no catálogo"
          },
          "slotBindings": {
            "$ref": "#/components/schemas/SlotBindingsDto"
          },
          "schedule": {
            "type": "object",
            "description": "Mapa dia-da-semana → horas UTC",
            "example": {
              "2": [
                18
              ],
              "3": [
                18
              ],
              "4": [
                13,
                18
              ]
            }
          },
          "categories": {
            "minItems": 0,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "leadsPerRun": {
            "type": "number",
            "minimum": 1,
            "description": "Leads contactados por execução (default 5)",
            "example": 5,
            "default": 5
          },
          "sendIntervalSeconds": {
            "type": "number",
            "minimum": 0,
            "description": "Intervalo em segundos entre envios (default 5)",
            "example": 5,
            "default": 5
          }
        }
      },
      "PatchOutreachConfigDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Habilitar outreach"
          },
          "outreachTemplateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de outreach no catálogo"
          },
          "notifyTemplateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de notify no catálogo"
          },
          "slotBindings": {
            "$ref": "#/components/schemas/SlotBindingsDto"
          },
          "schedule": {
            "type": "object",
            "description": "Mapa dia-da-semana → horas UTC",
            "example": {
              "2": [
                18
              ],
              "3": [
                18
              ],
              "4": [
                13,
                18
              ]
            }
          },
          "categories": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "leadsPerRun": {
            "type": "number",
            "minimum": 1,
            "description": "Leads contactados por execução",
            "example": 5
          },
          "sendIntervalSeconds": {
            "type": "number",
            "minimum": 0,
            "description": "Intervalo em segundos entre envios",
            "example": 5
          }
        }
      },
...
      "UpsertPlatformJobScheduleDto": {
        "type": "object",
        "properties": {
          "cronExpression": {
            "type": "string",
            "minLength": 1,
            "pattern": "FIVE_FIELD_CRON",
            "description": "Expressão cron de 5 campos",
            "example": "0 8 * * *"
          },
          "timeZone": {
            "type": "string",
            "minLength": 1,
            "description": "Timezone IANA",
            "default": "America/Sao_Paulo",
            "example": "America/Sao_Paulo"
          },
          "enabled": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "cronExpression"
        ]
      },
...
      "OpsSummaryDto": {
        "type": "object",
        "properties": {
          "totalTenants": {
            "type": "number",
            "example": 10
          },
          "activeTenants": {
            "type": "number",
            "example": 8
          },
          "outreachEnabledTenants": {
            "type": "number",
            "example": 3
          },
          "totalLeads": {
            "type": "number",
            "description": "Leads no pool global com deletedAt null",
            "example": 150
          }
        },
        "required": [
          "totalTenants",
          "activeTenants",
          "outreachEnabledTenants",
          "totalLeads"
        ]
      },
      "TenantLeadStatsDto": {
        "type": "object",
        "properties": {
          "contacted": {
            "type": "number",
            "example": 12
          },
          "replied": {
            "type": "number",
            "example": 4
          },
          "quoted": {
            "type": "number",
            "example": 1
          },
          "closed": {
            "type": "number",
            "example": 0
          },
          "deleted": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "contacted",
          "replied",
          "quoted",
          "closed",
          "deleted"
        ]
      },
      "LeadCountDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "description": "Quantidade de Lead com deletedAt null",
            "example": 150
          }
        },
        "required": [
          "count"
        ]
      },
...
    }
  }
}